# internal/bytealg: optimize indexbyte in amd64